Split arm and sve arch definitions
