+++ include/eve/arch/arm/sve/tags.hpp
@@ -16,11 +16,11 @@ namespace eve
   //================================================================================================
   // ABI tags for all ARM bits SIMD registers
   //================================================================================================
-  template<std::size_t Size, bool Logical> struct arm_abi_
+  template<std::size_t Size, bool Logical> struct sve_abi_
   {
-    static constexpr std::size_t bits                     = Size;
-    static constexpr std::size_t bytes                    = Size/8;
-    static constexpr bool        is_wide_logical = Logical;
+    static constexpr std::size_t bits             = Size;
+    static constexpr std::size_t bytes            = Size/8;
+    static constexpr bool        is_wide_logical  = Logical;
 
     template<typename Type>
     static constexpr bool is_full = ((Type::size() * sizeof(typename Type::value_type)) >= 8);
@@ -32,26 +32,16 @@ namespace eve
     static constexpr std::size_t fundamental_cardinal = 8 / sizeof(Type);
   };
 
-  struct arm_64_      : arm_abi_< 64,true>  {};
-  struct arm_128_     : arm_abi_<128,true>  {};
-  struct arm_sve_128_ : arm_abi_<128,false> {};
-  struct arm_sve_256_ : arm_abi_<256,false> {};
-  struct arm_sve_512_ : arm_abi_<512,false> {};
+  struct arm_sve_128_ : sve_abi_<128,false> {};
+  struct arm_sve_256_ : sve_abi_<256,false> {};
+  struct arm_sve_512_ : sve_abi_<512,false> {};
 
   //================================================================================================
   // Dispatching tag for ARM SIMD implementation
   //================================================================================================
-  struct neon128_ : simd_api<simd_   , spy::neon_>      {};
-  struct asimd_   : simd_api<neon128_, spy::asimd_>     {};
   struct sve128_  : simd_api<simd_   , spy::fixed_sve_> {};
   struct sve256_  : simd_api<simd_   , spy::fixed_sve_> {};
   struct sve512_  : simd_api<simd_   , spy::fixed_sve_> {};
-
-  //================================================================================================
-  // NEON extensions tag objects
-  //================================================================================================
-  inline constexpr neon128_ neon    = {};
-  inline constexpr asimd_   asimd   = {};
 
   //================================================================================================
   // SVE extensions tag objects
@@ -61,9 +51,8 @@ namespace eve
   inline constexpr sve512_  sve512  = {};
 
   //================================================================================================
-  // ARM ABI concept
+  // ARM SVE ABI concept
   //================================================================================================
-  template<typename T> concept arm_abi = detail::is_one_of<T> ( detail::types<arm_64_,arm_128_> {});
   template<typename T> concept sve_abi = detail::is_one_of<T> ( detail::types < arm_sve_128_
                                                                               , arm_sve_256_
                                                                               , arm_sve_512_
