+++ include/eve/arch/as_register.hpp
@@ -12,6 +12,7 @@
 #if !defined(EVE_NO_SIMD)
 #include <eve/arch/x86/as_register.hpp>
 #include <eve/arch/ppc/as_register.hpp>
-#include <eve/arch/arm/as_register.hpp>
+#include <eve/arch/arm/sve/as_register.hpp>
+#include <eve/arch/arm/neon/as_register.hpp>
 #endif
