+++ include/eve/arch/tags.hpp
@@ -10,4 +10,5 @@
 #include <eve/arch/cpu/tags.hpp>
 #include <eve/arch/x86/tags.hpp>
 #include <eve/arch/ppc/tags.hpp>
-#include <eve/arch/arm/tags.hpp>
+#include <eve/arch/arm/sve/tags.hpp>
+#include <eve/arch/arm/neon/tags.hpp>
